refactor: ディレクトリ名の変更
複数形の方が適切なため
