+++ frontend/src/components/pages/DivisionsNew.tsx
@@ -10,8 +10,8 @@ import {
   DivisionsNewResponse,
   User,
 } from "../../types";
-import { DeadlineTextField } from "../model/task/DeadlineTextField";
-import { PriorityTextField } from "../model/task/PriorityTextField";
+import { DeadlineTextField } from "../models/task/DeadlineTextField";
+import { PriorityTextField } from "../models/task/PriorityTextField";
 import { SnackbarContext } from "../../providers/SnackbarProvider";
 
 const DivisionsNew: FC = () => {
