+++ frontend/src/components/pages/TasksEdit.tsx
@@ -7,8 +7,8 @@ import { axiosInstance } from "../../utils/axios";
 import { TasksResponse, EditTask } from "../../types";
 import { AuthContext } from "../../providers/AuthProvider";
 import { useFetchTask } from "../../hooks/useFetchTask";
-import { PriorityTextField } from "../model/task/PriorityTextField";
-import { DeadlineTextField } from "../model/task/DeadlineTextField";
+import { PriorityTextField } from "../models/task/PriorityTextField";
+import { DeadlineTextField } from "../models/task/DeadlineTextField";
 import { SnackbarContext } from "../../providers/SnackbarProvider";
 
 const TasksEdit: FC = () => {
